new code

+++ app_pojavlauncher/src/main/java/net/kdt/pojavlaunch/JAssetInfo.kt
@@ -5,5 +5,5 @@ import androidx.annotation.Keep
 @Keep
 class JAssetInfo {
     var hash: String? = null
-    var size: Int = 0
+    var size: Long = 0
 }
